Rename sampler -> recorder
