+++ src/recorder/snapshot.rs
